ROX-36586: Move validation schema to same folder as wizard step 2
