+++ ui/apps/platform/src/Containers/Vulnerabilities/Reports/ImageVulnerabilityReports/View/ImageVulnerabilityFiltersView.tsx
@@ -15,9 +15,9 @@ import type {
     CompoundSearchFilterAttribute,
     CompoundSearchFilterConfig,
 } from 'Components/CompoundSearchFilter/types';
+import type { ImageVulnerabilityReportFiltersConfigurationForEntity } from 'services/ReportsService.types';
 import { getSearchFilterFromSearchString } from 'utils/searchUtils';
 
-import type { ImageVulnerabilityFiltersConfigurationForEntity } from '../imageVulnerabilityReports.types';
 import { getTextForCvesSince } from '../imageVulnerabilityReports.utils';
 
 export type ImageVulnerabilityFiltersViewProps = {
@@ -25,7 +25,7 @@ export type ImageVulnerabilityFiltersViewProps = {
     headingLevel: 'h2' | 'h3';
     horizontalTermWidthModifier: BreakpointModifiers;
     searchFilterConfig: CompoundSearchFilterConfig;
-    values: ImageVulnerabilityFiltersConfigurationForEntity;
+    values: ImageVulnerabilityReportFiltersConfigurationForEntity;
 };
 
 function ImageVulnerabilityFiltersView({
