+++ ui/apps/platform/src/Containers/Vulnerabilities/Reports/ImageVulnerabilityReports/View/ImageVulnerabilityFiltersViewForCollection.tsx
@@ -12,14 +12,14 @@ import type { BreakpointModifiers } from '@patternfly/react-core';
 
 import { fixabilityLabels } from 'constants/reportConstants';
 import { vulnerabilitySeverityLabels } from 'messages/common';
+import type { ImageVulnerabilityReportFiltersConfigurationForCollection } from 'services/ReportsService.types';
 
-import type { ImageVulnerabilityFiltersConfigurationForCollection } from '../imageVulnerabilityReports.types';
 import { getTextForCvesSince } from '../imageVulnerabilityReports.utils';
 
 export type ImageVulnerabilityFiltersViewForCollectionProps = {
     headingLevel: 'h2' | 'h3';
     horizontalTermWidthModifier: BreakpointModifiers;
-    values: ImageVulnerabilityFiltersConfigurationForCollection;
+    values: ImageVulnerabilityReportFiltersConfigurationForCollection;
 };
 
 function ImageVulnerabilityFiltersViewForCollection({
