+++ ui/apps/platform/src/Containers/Vulnerabilities/Reports/ImageVulnerabilityReports/View/ImageVulnerabilityResourcesView.tsx
@@ -11,14 +11,14 @@ import {
 import type { BreakpointModifiers } from '@patternfly/react-core';
 
 import EntityScopeDescriptionListGroups from 'Components/EntityScope/EntityScopeDescriptionListGroups';
+import type { ImageVulnerabilityReportResourcesConfiguration } from 'services/ReportsService.types';
 
-import type { ImageVulnerabilityResourcesConfiguration } from '../imageVulnerabilityReports.types';
 import { imageTypeLabelMap } from '../imageVulnerabilityReports.utils';
 
 export type ImageVulnerabilityResourcesViewProps = {
     headingLevel: 'h2' | 'h3';
     horizontalTermWidthModifier: BreakpointModifiers;
-    values: ImageVulnerabilityResourcesConfiguration;
+    values: ImageVulnerabilityReportResourcesConfiguration;
 };
 
 function ImageVulnerabilityResourcesView({
