+++ ui/apps/platform/src/Containers/Vulnerabilities/Reports/ImageVulnerabilityReports/Wizard/ImageVulnerabilityReportWizardPage.tsx
@@ -44,8 +44,7 @@ import {
     searchFilterConfigForImageVulnerabilityReport,
 } from '../../../searchFilterConfig';
 
-import { defaultImageVulnerabilityReportConfiguration } from '../imageVulnerabilityReports.defaults';
-
+import { defaultImageVulnerabilityReportConfiguration } from './imageVulnerabilityReports.defaults';
 import ImageVulnerabilityReportWizard from './ImageVulnerabilityReportWizard';
 
 const searchFilterLabelsForImages: SearchFieldLabel[] = [
