+++ ui/apps/platform/src/Containers/Vulnerabilities/Reports/ImageVulnerabilityReports/Wizard/Step2/ImageVulnerabilityResourcesStep.tsx
@@ -15,12 +15,10 @@ import type {
     EntityScope,
     EntityScopeRule,
     ImageType,
+    ImageVulnerabilityReportFiltersConfiguration,
+    ImageVulnerabilityReportResourcesConfiguration,
 } from 'services/ReportsService.types';
 
-import type {
-    ImageVulnerabilityFiltersConfiguration,
-    ImageVulnerabilityResourcesConfiguration,
-} from '../../imageVulnerabilityReports.types';
 import {
     replaceFiltersForCollectionFromQueryInConfiguration,
     replaceFiltersForEntityFromFixabilityAndSeverityInConfiguration,
@@ -34,13 +32,15 @@ function hasWatchedImagesOnly(imageTypes: ImageType[]) {
 }
 
 export type ImageVulnerabilityResourcesStepProps<
-    T extends ImageVulnerabilityResourcesConfiguration = ImageVulnerabilityResourcesConfiguration,
+    T extends ImageVulnerabilityReportResourcesConfiguration =
+        ImageVulnerabilityReportResourcesConfiguration,
 > = {
     formik: FormikProps<T>;
 };
 
 function ImageVulnerabilityResourcesStep<
-    T extends ImageVulnerabilityResourcesConfiguration = ImageVulnerabilityResourcesConfiguration,
+    T extends ImageVulnerabilityReportResourcesConfiguration =
+        ImageVulnerabilityReportResourcesConfiguration,
 >({ formik }: ImageVulnerabilityResourcesStepProps<T>): ReactElement {
     const [collectionScopeSelected, setCollectionScopeSelected] = useState<CollectionScope>(
         'collectionScope' in formik.values.resourceScope
@@ -57,11 +57,11 @@ function ImageVulnerabilityResourcesStep<
     // If user selects entityScope again, then restore it instead of having forgotten it.
     // Forgive double type assertion until we delete deprecated collections.
     const [vulnReportFiltersForEntityScope, setVulnReportFiltersForEntityScope] = useState<
-        ImageVulnerabilityResourcesConfiguration['vulnReportFilters']
+        ImageVulnerabilityReportResourcesConfiguration['vulnReportFilters']
     >(
         replaceFiltersForEntityFromFixabilityAndSeverityInConfiguration(
-            formik.values as unknown as ImageVulnerabilityFiltersConfiguration
-        ) as unknown as ImageVulnerabilityResourcesConfiguration['vulnReportFilters']
+            formik.values as unknown as ImageVulnerabilityReportFiltersConfiguration
+        ) as unknown as ImageVulnerabilityReportResourcesConfiguration['vulnReportFilters']
     );
 
     // Remember rules property of entityScope if user selects WATCHED only.
@@ -103,7 +103,7 @@ function ImageVulnerabilityResourcesStep<
         formik.setFieldValue(
             'vulnReportFilters',
             replaceFiltersForCollectionFromQueryInConfiguration(
-                formik.values as unknown as ImageVulnerabilityFiltersConfiguration
+                formik.values as unknown as ImageVulnerabilityReportFiltersConfiguration
             )
         );
     }
