+++ ui/apps/platform/src/Containers/Vulnerabilities/Reports/ImageVulnerabilityReports/Wizard/Step2/validationSchemaResources.ts
@@ -1,0 +1,45 @@
+import * as yup from 'yup';
+
+import { imageTypes } from 'services/ReportsService.types';
+import type {
+    ImageType,
+    ImageVulnerabilityReportResourceScope,
+    ImageVulnerabilityReportResourcesConfiguration,
+} from 'services/ReportsService.types';
+
+export const validationSchemaResources: yup.ObjectSchema<ImageVulnerabilityReportResourcesConfiguration> =
+    yup.object().shape({
+        vulnReportFilters: yup.object().shape({
+            imageTypes: yup
+                .array()
+                .of(yup.string().oneOf(imageTypes).required())
+                .min(1, 'Select at least one image type')
+                .required(),
+        }),
+        resourceScope: yup
+            .object()
+            .test(
+                'resource-scope-required',
+                'Resource scope is required for deployed images',
+                function test(value) {
+                    // resourceScope can be either CollectionScope or EntityScope, which .shape()
+                    // can't express. Cast so property access is type checked.
+                    const resourceScope = value as ImageVulnerabilityReportResourceScope;
+                    const imageTypesValue: ImageType[] =
+                        this.parent?.vulnReportFilters?.imageTypes ?? [];
+                    const hasDeployed = imageTypesValue.includes('DEPLOYED');
+
+                    if (!hasDeployed) {
+                        return true;
+                    }
+
+                    if ('collectionScope' in resourceScope) {
+                        return Boolean(resourceScope.collectionScope.collectionId);
+                    }
+                    if ('entityScope' in resourceScope) {
+                        return resourceScope.entityScope.rules.length > 0;
+                    }
+                    return false;
+                }
+            ) as unknown as yup.Schema<ImageVulnerabilityReportResourceScope>,
+    });
