+++ ui/apps/platform/src/Containers/Vulnerabilities/Reports/ImageVulnerabilityReports/Wizard/Step3/ImageVulnerabilityFiltersSince.tsx
@@ -5,21 +5,23 @@ import type { FormikProps } from 'formik';
 
 import FormLabelGroup from 'Components/PatternFly/FormLabelGroup';
 import SelectSingle from 'Components/SelectSingle/SelectSingle';
+import type { ImageVulnerabilityReportFiltersConfiguration } from 'services/ReportsService.types';
 
-import type { ImageVulnerabilityFiltersConfiguration } from '../../imageVulnerabilityReports.types';
 import {
     getReportFiltersWithoutCvesSince,
     getValueForCvesSince,
 } from '../../imageVulnerabilityReports.utils';
 
 export type ImageVulnerabilityFiltersSinceProps<
-    T extends ImageVulnerabilityFiltersConfiguration = ImageVulnerabilityFiltersConfiguration,
+    T extends ImageVulnerabilityReportFiltersConfiguration =
+        ImageVulnerabilityReportFiltersConfiguration,
 > = {
     formik: FormikProps<T>;
 };
 
 function ImageVulnerabilityFiltersSince<
-    T extends ImageVulnerabilityFiltersConfiguration = ImageVulnerabilityFiltersConfiguration,
+    T extends ImageVulnerabilityReportFiltersConfiguration =
+        ImageVulnerabilityReportFiltersConfiguration,
 >({ formik }: ImageVulnerabilityFiltersSinceProps<T>): ReactElement {
     // Initialize with either selected start date or unspecified value.
     // Objective of local state is remember date in scenario:
