+++ ui/apps/platform/src/Containers/Vulnerabilities/Reports/ImageVulnerabilityReports/Wizard/Step3/ImageVulnerabilityFiltersStep.tsx
@@ -9,14 +9,13 @@ import type {
 } from 'Components/CompoundSearchFilter/types';
 
 import FiltersQuery from 'Components/Reports/Step/FiltersQuery';
-
-import type { ImageVulnerabilityFiltersConfigurationForEntity } from '../../imageVulnerabilityReports.types';
+import type { ImageVulnerabilityReportFiltersConfigurationForEntity } from 'services/ReportsService.types';
 
 import ImageVulnerabilityFiltersSince from './ImageVulnerabilityFiltersSince';
 
 export type ImageVulnerabilityFiltersStepProps<
-    T extends ImageVulnerabilityFiltersConfigurationForEntity =
-        ImageVulnerabilityFiltersConfigurationForEntity,
+    T extends ImageVulnerabilityReportFiltersConfigurationForEntity =
+        ImageVulnerabilityReportFiltersConfigurationForEntity,
 > = {
     attributesSeparateFromConfig: SelectSearchFilterAttribute[];
     formik: FormikProps<T>;
@@ -24,8 +23,8 @@ export type ImageVulnerabilityFiltersStepProps<
 };
 
 function ImageVulnerabilityFiltersStep<
-    T extends ImageVulnerabilityFiltersConfigurationForEntity =
-        ImageVulnerabilityFiltersConfigurationForEntity,
+    T extends ImageVulnerabilityReportFiltersConfigurationForEntity =
+        ImageVulnerabilityReportFiltersConfigurationForEntity,
 >({
     attributesSeparateFromConfig,
     formik,
