+++ ui/apps/platform/src/Containers/Vulnerabilities/Reports/ImageVulnerabilityReports/Wizard/Step3/ImageVulnerabilityFiltersStepForCollection.tsx
@@ -7,9 +7,9 @@ import SearchFilterSelectInclusive from 'Components/CompoundSearchFilter/compone
 import FormLabelGroup from 'Components/PatternFly/FormLabelGroup';
 import type { OnSearchPayload } from 'Components/CompoundSearchFilter/types';
 import { updateSearchFilter } from 'Components/CompoundSearchFilter/utils/utils';
+import type { ImageVulnerabilityReportFiltersConfigurationForCollection } from 'services/ReportsService.types';
 
 import { attributeForSeverityInBackendAndViewBasedReport } from '../../../../searchFilterConfig';
-import type { ImageVulnerabilityFiltersConfigurationForCollection } from '../../imageVulnerabilityReports.types';
 import {
     attributeForFixabilityInFiltersStepForCollection,
     getFixabilityFromSearchFilter,
@@ -20,15 +20,15 @@ import {
 import ImageVulnerabilityFiltersSince from './ImageVulnerabilityFiltersSince';
 
 export type ImageVulnerabilityFiltersStepForCollectionProps<
-    T extends ImageVulnerabilityFiltersConfigurationForCollection =
-        ImageVulnerabilityFiltersConfigurationForCollection,
+    T extends ImageVulnerabilityReportFiltersConfigurationForCollection =
+        ImageVulnerabilityReportFiltersConfigurationForCollection,
 > = {
     formik: FormikProps<T>;
 };
 
 function ImageVulnerabilityFiltersStepForCollection<
-    T extends ImageVulnerabilityFiltersConfigurationForCollection =
-        ImageVulnerabilityFiltersConfigurationForCollection,
+    T extends ImageVulnerabilityReportFiltersConfigurationForCollection =
+        ImageVulnerabilityReportFiltersConfigurationForCollection,
 >({ formik }: ImageVulnerabilityFiltersStepForCollectionProps<T>): ReactElement {
     const searchFilter = getSearchFilterFromReportConfiguration(formik.values.vulnReportFilters);
