+++ ui/apps/platform/src/Containers/Vulnerabilities/Reports/ImageVulnerabilityReports/Wizard/Step3/validationSchemaFilters.ts
@@ -1,0 +1,89 @@
+import * as yup from 'yup';
+import { isValid, parse } from 'date-fns';
+
+import type {
+    CvesSince,
+    ImageVulnerabilityReportConfiguration,
+    ImageVulnerabilityReportConfigurationForCollection,
+    ImageVulnerabilityReportFiltersForEntity,
+} from 'services/ReportsService.types';
+import { vulnerabilitySeverities } from 'types/cve.proto';
+
+function isCvesSinceValid(filters: CvesSince): boolean {
+    if ('allVuln' in filters) {
+        return true;
+    }
+    if ('sinceLastSentScheduledReport' in filters) {
+        return true;
+    }
+    if ('sinceStartDate' in filters) {
+        if (!filters.sinceStartDate) {
+            return false;
+        }
+        // parse() is lenient. partial strings like "2026-10" pass isValid.
+        // Checking for "T" ensures the value is likely a full ISO timestamp.
+        return filters.sinceStartDate.includes('T') && isValid(parse(filters.sinceStartDate));
+    }
+    return false;
+}
+
+export const validationSchemaFilters = yup.lazy((values: ImageVulnerabilityReportConfiguration) => {
+    if ('entityScope' in values.resourceScope) {
+        return yup.object().shape({
+            vulnReportFilters: yup
+                .object()
+                .shape({
+                    query: yup.string().required('At least one filter is required'),
+                })
+                .test(
+                    'cves-since-entity',
+                    'CVE date range is required',
+                    function test(reportFilters) {
+                        // CvesSince is a union type that .shape() can't express.
+                        // cast so property access is type checked.
+                        const filters = reportFilters as ImageVulnerabilityReportFiltersForEntity;
+                        if ('sinceStartDate' in filters && !isCvesSinceValid(filters)) {
+                            return this.createError({
+                                path: 'vulnReportFilters.sinceStartDate',
+                                message: 'A valid start date is required',
+                            });
+                        }
+                        return isCvesSinceValid(filters);
+                    }
+                ),
+        });
+    }
+
+    return yup.object().shape({
+        vulnReportFilters: yup
+            .object()
+            .shape({
+                severities: yup
+                    .array()
+                    .of(yup.string().oneOf(vulnerabilitySeverities).required())
+                    .min(1, 'Select at least one CVE severity')
+                    .required('CVE severity is required'),
+                fixability: yup
+                    .string()
+                    .oneOf(['FIXABLE', 'NOT_FIXABLE', 'BOTH'])
+                    .required('CVE status is required'),
+            })
+            .test(
+                'cves-since-collection',
+                'CVE date range is required',
+                function test(reportFilters) {
+                    // CvesSince is a union type that .shape() can't express.
+                    // cast so property access is type checked.
+                    const filters =
+                        reportFilters as ImageVulnerabilityReportConfigurationForCollection['vulnReportFilters'];
+                    if ('sinceStartDate' in filters && !isCvesSinceValid(filters)) {
+                        return this.createError({
+                            path: 'vulnReportFilters.sinceStartDate',
+                            message: 'A valid start date is required',
+                        });
+                    }
+                    return isCvesSinceValid(filters);
+                }
+            ),
+    });
+});
