+++ ui/apps/platform/src/Containers/Vulnerabilities/Reports/ImageVulnerabilityReports/Wizard/imageVulnerabilityReportValidationSchemas.ts
@@ -1,195 +1,11 @@
-import { isValid, parse } from 'date-fns';
 import * as yup from 'yup';
 
-import type { DeliveryType, DetailsType } from 'Components/Reports/reports.types';
-import type { Schedule } from 'types/schedule.proto';
-import { imageTypes } from 'services/ReportsService.types';
-import type {
-    CvesSince,
-    ImageType,
-    ImageVulnerabilityReportConfiguration,
-    ImageVulnerabilityReportConfigurationForCollection,
-    ImageVulnerabilityReportFiltersForEntity,
-} from 'services/ReportsService.types';
-import { vulnerabilitySeverities } from 'types/cve.proto';
-
+import { validationSchemaDelivery } from 'Components/Reports/Step/validationSchemaDelivery';
+import { validationSchemaDetails } from 'Components/Reports/Step/validationSchemaDetails';
 import { ensureExhaustive } from 'utils/type.utils';
 
-import type {
-    ImageVulnerabilityResourceScope,
-    ImageVulnerabilityResourcesConfiguration,
-} from '../imageVulnerabilityReports.types';
-
-const validationSchemaDetails: yup.ObjectSchema<DetailsType> = yup.object().shape({
-    name: yup.string().trim().required('Report name is required'),
-    description: yup.string().ensure(),
-});
-
-const validationSchemaResources: yup.ObjectSchema<ImageVulnerabilityResourcesConfiguration> = yup
-    .object()
-    .shape({
-        vulnReportFilters: yup.object().shape({
-            imageTypes: yup
-                .array()
-                .of(yup.string().oneOf(imageTypes).required())
-                .min(1, 'Select at least one image type')
-                .required(),
-        }),
-        resourceScope: yup
-            .object()
-            .test(
-                'resource-scope-required',
-                'Resource scope is required for deployed images',
-                function test(value) {
-                    // resourceScope can be either CollectionScope or EntityScope, which .shape()
-                    // can't express. Cast so property access is type checked.
-                    const resourceScope = value as ImageVulnerabilityResourceScope;
-                    const imageTypesValue: ImageType[] =
-                        this.parent?.vulnReportFilters?.imageTypes ?? [];
-                    const hasDeployed = imageTypesValue.includes('DEPLOYED');
-
-                    if (!hasDeployed) {
-                        return true;
-                    }
-
-                    if ('collectionScope' in resourceScope) {
-                        return Boolean(resourceScope.collectionScope.collectionId);
-                    }
-                    if ('entityScope' in resourceScope) {
-                        return resourceScope.entityScope.rules.length > 0;
-                    }
-                    return false;
-                }
-            ) as unknown as yup.Schema<ImageVulnerabilityResourceScope>,
-    });
-
-function isCvesSinceValid(filters: CvesSince): boolean {
-    if ('allVuln' in filters) {
-        return true;
-    }
-    if ('sinceLastSentScheduledReport' in filters) {
-        return true;
-    }
-    if ('sinceStartDate' in filters) {
-        if (!filters.sinceStartDate) {
-            return false;
-        }
-        // parse() is lenient. partial strings like "2026-10" pass isValid.
-        // Checking for "T" ensures the value is likely a full ISO timestamp.
-        return filters.sinceStartDate.includes('T') && isValid(parse(filters.sinceStartDate));
-    }
-    return false;
-}
-
-const validationSchemaFilters = yup.lazy((values: ImageVulnerabilityReportConfiguration) => {
-    if ('entityScope' in values.resourceScope) {
-        return yup.object().shape({
-            vulnReportFilters: yup
-                .object()
-                .shape({
-                    query: yup.string().required('At least one filter is required'),
-                })
-                .test(
-                    'cves-since-entity',
-                    'CVE date range is required',
-                    function test(reportFilters) {
-                        // CvesSince is a union type that .shape() can't express.
-                        // cast so property access is type checked.
-                        const filters = reportFilters as ImageVulnerabilityReportFiltersForEntity;
-                        if ('sinceStartDate' in filters && !isCvesSinceValid(filters)) {
-                            return this.createError({
-                                path: 'vulnReportFilters.sinceStartDate',
-                                message: 'A valid start date is required',
-                            });
-                        }
-                        return isCvesSinceValid(filters);
-                    }
-                ),
-        });
-    }
-
-    return yup.object().shape({
-        vulnReportFilters: yup
-            .object()
-            .shape({
-                severities: yup
-                    .array()
-                    .of(yup.string().oneOf(vulnerabilitySeverities).required())
-                    .min(1, 'Select at least one CVE severity')
-                    .required('CVE severity is required'),
-                fixability: yup
-                    .string()
-                    .oneOf(['FIXABLE', 'NOT_FIXABLE', 'BOTH'])
-                    .required('CVE status is required'),
-            })
-            .test(
-                'cves-since-collection',
-                'CVE date range is required',
-                function test(reportFilters) {
-                    // CvesSince is a union type that .shape() can't express.
-                    // cast so property access is type checked.
-                    const filters =
-                        reportFilters as ImageVulnerabilityReportConfigurationForCollection['vulnReportFilters'];
-                    if ('sinceStartDate' in filters && !isCvesSinceValid(filters)) {
-                        return this.createError({
-                            path: 'vulnReportFilters.sinceStartDate',
-                            message: 'A valid start date is required',
-                        });
-                    }
-                    return isCvesSinceValid(filters);
-                }
-            ),
-    });
-});
-
-const validationSchemaDelivery: yup.ObjectSchema<DeliveryType> = yup.object().shape({
-    notifiers: yup
-        .array()
-        .of(
-            yup.object().shape({
-                emailConfig: yup.object().shape({
-                    notifierId: yup.string().required('Email notifier is required'),
-                    mailingLists: yup
-                        .array()
-                        .of(yup.string().required())
-                        .min(1, 'At least one distribution list email is required')
-                        .required(),
-                    customSubject: yup.string().defined(),
-                    customBody: yup.string().defined(),
-                }),
-                notifierName: yup.string().defined(),
-            })
-        )
-        .defined(),
-    schedule: yup
-        .object()
-        .nullable()
-        .defined()
-        .test('schedule-days-required', 'At least one day must be selected', function test(value) {
-            if (value == null) {
-                return true;
-            }
-            // Schedule is a union type that .shape() can't express.
-            // Cast so property access is type-checked.
-            const schedule = value as Schedule;
-            if (schedule.intervalType === 'UNSET' || schedule.intervalType === 'DAILY') {
-                return true;
-            }
-            if ('daysOfWeek' in schedule && schedule.daysOfWeek.days.length === 0) {
-                return this.createError({
-                    path: 'schedule.daysOfWeek.days',
-                    message: 'At least one day must be selected',
-                });
-            }
-            if ('daysOfMonth' in schedule && schedule.daysOfMonth.days.length === 0) {
-                return this.createError({
-                    path: 'schedule.daysOfMonth.days',
-                    message: 'At least one day must be selected',
-                });
-            }
-            return true;
-        }) as unknown as yup.Schema<Schedule | null>,
-});
+import { validationSchemaResources } from './Step2/validationSchemaResources';
+import { validationSchemaFilters } from './Step3/validationSchemaFilters';
 
 export const wizardStepIds = ['details', 'resources', 'filters', 'delivery', 'review'] as const;
 export type WizardStepId = (typeof wizardStepIds)[number];
