+++ ui/apps/platform/src/Containers/Vulnerabilities/Reports/ImageVulnerabilityReports/Wizard/imageVulnerabilityReports.defaults.ts
@@ -4,8 +4,8 @@ import type {
 } from 'services/ReportsService.types';
 import { getRequestQueryStringForSearchFilter } from 'utils/searchUtils';
 
-import { defaultStorage } from '../../WorkloadCves/Overview/WorkloadCvesOverviewPage';
-import { fixableStatusToFixability, severityLabelToSeverity } from '../../utils/searchUtils';
+import { defaultStorage } from '../../../WorkloadCves/Overview/WorkloadCvesOverviewPage';
+import { fixableStatusToFixability, severityLabelToSeverity } from '../../../utils/searchUtils';
 
 const {
     preferences: { defaultFilters },
