--- ui/apps/platform/src/Containers/Vulnerabilities/Reports/ImageVulnerabilityReports/imageVulnerabilityReports.types.ts
@@ -1,57 +1,0 @@
-import type {
-    CollectionScope,
-    CvesSince,
-    EntityScope,
-    Fixability,
-    ImageType,
-    NotifierConfiguration,
-} from 'services/ReportsService.types';
-import type { VulnerabilitySeverity } from 'types/cve.proto';
-import type { Schedule } from 'types/schedule.proto';
-
-// Types because Resources and Filters for view and wizard combine data properties.
-
-export type ImageVulnerabilityResourceScope =
-    | {
-          collectionScope: CollectionScope;
-      }
-    | {
-          entityScope: EntityScope;
-      };
-
-export type ImageVulnerabilityResourcesConfiguration = {
-    resourceScope: ImageVulnerabilityResourceScope;
-    vulnReportFilters: {
-        imageTypes: ImageType[]; // more closely related to resources although in vulnReportFilters
-    };
-};
-
-export type ImageVulnerabilityFiltersConfiguration = {
-    vulnReportFilters: ImageVulnerabilityFiltersWithoutCvesSince & CvesSince;
-};
-
-export type ImageVulnerabilityFiltersWithoutCvesSince =
-    | ImageVulnerabilityFiltersForEntityWithoutCvesSince
-    | ImageVulnerabilityFiltersForCollectionWithoutCvesSince;
-
-export type ImageVulnerabilityFiltersConfigurationForEntity = {
-    vulnReportFilters: ImageVulnerabilityFiltersForEntityWithoutCvesSince & CvesSince;
-};
-
-export type ImageVulnerabilityFiltersForEntityWithoutCvesSince = {
-    query: string;
-};
-
-export type ImageVulnerabilityFiltersConfigurationForCollection = {
-    vulnReportFilters: ImageVulnerabilityFiltersForCollectionWithoutCvesSince & CvesSince;
-};
-
-export type ImageVulnerabilityFiltersForCollectionWithoutCvesSince = {
-    fixability: Fixability;
-    severities: VulnerabilitySeverity[];
-};
-
-export type DestinationsConfiguration = {
-    notifiers: NotifierConfiguration[];
-    schedule: Schedule | null;
-};
