+++ ui/apps/platform/src/Containers/Vulnerabilities/Reports/ImageVulnerabilityReports/imageVulnerabilityReports.utils.ts
@@ -6,6 +6,9 @@ import type {
     ImageVulnerabilityReportConfiguration,
     ImageVulnerabilityReportConfigurationForCollection,
     ImageVulnerabilityReportConfigurationForEntity,
+    ImageVulnerabilityReportFiltersConfiguration,
+    ImageVulnerabilityReportFiltersForCollectionWithoutCvesSince,
+    ImageVulnerabilityReportFiltersWithoutCvesSince,
 } from 'services/ReportsService.types';
 import type { VulnerabilitySeverity } from 'types/cve.proto';
 import type { SearchFilter } from 'types/search';
@@ -17,12 +20,6 @@ import {
 } from 'utils/searchUtils';
 import { ensureExhaustive } from 'utils/type.utils';
 import { vulnerabilityConfigurationsReportsPath } from 'routePaths';
-
-import type {
-    ImageVulnerabilityFiltersConfiguration,
-    ImageVulnerabilityFiltersForCollectionWithoutCvesSince,
-    ImageVulnerabilityFiltersWithoutCvesSince,
-} from './imageVulnerabilityReports.types';
 
 // TypeScript needs help because entityScope and vulnReportFilters have mutual dependencies.
 // entityScope implies query but not fixability nore severities
@@ -69,8 +66,8 @@ export function getValueForCvesSince(vulnReportFilters: CvesSince) {
 // For future entity scope only, replace helper function call with the following:
 // const { query } = formik.values.vulnReportFilters;
 export function getReportFiltersWithoutCvesSince(
-    values: ImageVulnerabilityFiltersConfiguration
-): ImageVulnerabilityFiltersWithoutCvesSince {
+    values: ImageVulnerabilityReportFiltersConfiguration
+): ImageVulnerabilityReportFiltersWithoutCvesSince {
     // Idiom to ignore properties is awkward because of disjunction.
     /* eslint-disable @typescript-eslint/no-unused-vars */
     // eslint-diable-next-line @typescript-eslint/ban-ts-comment
@@ -114,7 +111,8 @@ function getFixableArrayFromFixability(fixability: Fixability): string[] {
 // TypeScript needs if statement and default return statement, but code does not call in that scenario.
 export function replaceFiltersForCollectionFromQueryInConfiguration({
     vulnReportFilters,
-}: ImageVulnerabilityFiltersConfiguration): ImageVulnerabilityFiltersWithoutCvesSince & CvesSince {
+}: ImageVulnerabilityReportFiltersConfiguration): ImageVulnerabilityReportFiltersWithoutCvesSince &
+    CvesSince {
     if ('query' in vulnReportFilters) {
         const { query, ...restOfFilters } = vulnReportFilters;
         const searchFilter = getSearchFilterFromSearchString(query);
@@ -133,7 +131,8 @@ export function replaceFiltersForCollectionFromQueryInConfiguration({
 // otherwise return the original vulnReportFilters object for entityScope.
 export function replaceFiltersForEntityFromFixabilityAndSeverityInConfiguration({
     vulnReportFilters,
-}: ImageVulnerabilityFiltersConfiguration): ImageVulnerabilityFiltersWithoutCvesSince & CvesSince {
+}: ImageVulnerabilityReportFiltersConfiguration): ImageVulnerabilityReportFiltersWithoutCvesSince &
+    CvesSince {
     if ('fixability' in vulnReportFilters && 'severities' in vulnReportFilters) {
         const { fixability, severities, ...restOfFilters } = vulnReportFilters;
 
@@ -167,7 +166,7 @@ export const attributeForFixabilityInFiltersStepForCollection: SelectExclusiveSi
 
 // Pseudo search filter (for select elements) corresponds to property names and preceding pseudo attribute.
 export function getSearchFilterFromReportConfiguration(
-    vulnReportFilters: ImageVulnerabilityFiltersForCollectionWithoutCvesSince
+    vulnReportFilters: ImageVulnerabilityReportFiltersForCollectionWithoutCvesSince
 ): SearchFilter {
     return {
         Fixable: vulnReportFilters.fixability,
